Set timeout for EPG events to 1 second (LP: #434994)

+++ src/dvb_scanner.cc
@@ -73,7 +73,7 @@ void Scanner::tune_to(Frontend& frontend, const Transponder& transponder)
 				transponder.polarisation);
 		}
 
-		g_debug("Got %d transponders from NIT", nis.transponders.size());
+		g_debug("Got %u transponders from NIT", (guint)nis.transponders.size());
 		for (guint i = 0; i < nis.transponders.size(); i++)
 		{
 			Transponder& new_transponder = nis.transponders[i];
